Print help message when put wrong command (#32)
To improve the command line help message, whenever input wrong command or option, it will show help message. 
if the command is wrong, it will show help message,  exact the same as "moneyd --help"
if the command is correct, but option is wrong, it will show help message with corresponding options to that command.